Event Controller: Set propagation phase to capture (#223)
Fixes https://github.com/workbenchdev/demos/issues/222


button with "suggested-action" style

The reason that happens is that when clicking on a button, the button
claims the "released" gesture
and we don't get the "released" signal.
Setting the propagation phase to "CAPTURE" allows
us to handle the event before the button claims
it.